Added a angle
The scrypt now checks if the thrusters are facing the correct way befor applying thrust. Currently it is binary checking so thrust is On or Off, but this should be modified to be more like a porabla formula so thrust will be applied even when the trusters arent fully lined up. However things are MUCH smoother now for large ships.